HOMEに戻るとバグるやつを解消
